Make TagChecks check arbitrary boolean expressions on tag presence. (#173)
* Make TagChecks check arbitrary boolean expressions on tag presence.

This commit updates TagCheck to refer to a `Predicate` rather than the
simple presence of a tag. This allows us to express arbitrary boolean
expressions involving the presence of tags on an access path.

Some changes had to be made pretty far away from `TagCheck` to make this
happen. In particular, our desire to keep `TagChecks` copiable for now but
to have a clear single owner for `Predicate`s caused us to create
a special `PredicateDecoder` class which just governs access to the code
that creates `Predicate`s from protos and owns the unique pointers thus
generated. This, in turn, is held by a `ParticleSpec` to ensure that
`Predicate`s live at least as long as the `ParticleSpec` to ensure that we
don't delete `Predicate`s before the datalog has been printed out. I had
to make various accomodations to deal with these classes no longer being
copiable due to the `unique_ptr`s that they now held.

This change also affects the Datalog strings generated from a `TagCheck`
pretty dramatically. As a stopgap measure, @bgogul had created
`checkHasTag`, which took an access path and a tag as its arguments. Now
that a check is no longer completely summarized by an access path and
a tag, that had to change. This commit replaces `checkHasTag` with two
Datalog relations: `isCheck` and `check`, each of which takes
a label unique to the printed `TagCheck`. `isCheck` is an uncoditional
fact that puts the label in the `isCheck` universe. `check` acts as the
head of a rule with the TagCheck predicate as its body. With both of
these relations used as output relations, a driver can check that all
checks passed by ensuring that `isCheck` and `check` contain the same
contents.

Because we now have to generate unique labels, we introduce
a `DatalogPrintContext` to track mutable context that must be updated
during the printing process and use this to print our labels.

This commit fixes #138.

* Fix inaccurate comment about movability of `DatalogPrintContext`.

* Add comment explaining (non)ownership of `Predicate` on `TagCheck`.