A design pattern for non-unanimous decisions
This shows a policy where a non-unanimous decision is made. It is a
contrived example rather than a particular application. It is simplified
from [PR #141](https://github.com/google-research/raksha/pull/141),
which includes a larger application. The particular choice of dispute
resolution policy is now based on whether a principal occupies a
particular role.