A variant of the multimic policy.
This is an example of the multimic policy using a combination of IFC and
auth logic using the may/will check. Because in this scenario, a lot of
the information needed to make policy decisions is not available until
runtime, this example assumes that there is an interpreter and a runtime
that uses the interpreter to decide if particles can run (or not). (We
do not have an interpreter or a reference monitor, yet).