Multimic scenario, strongly tying policies to graph fragments.
Following a conversation with @bgogul, this attempt at the multimic
policy attempts to avoid expensive re-analysis of graph properties and
dynamic enforcement of policy by assuming that policies are strongly
associated with a particular graph fragment, within which the
connections cannot change dynamically, and that policies can speak only
of properties within those graph fragments.

This attempt introduces user permission stubs to the multimic graph.
These permission stubs represent the smart mic attempting to accomodate
user policies by describing the speech data as flowing through the
smartphones of guest users. Of course, the actual data does no such
thing, but this allows the smartphones to add tags to the dataflow while
also restricting the particles that they can control to the particles
that they own, allowing the microphone's actual graph to not be altered
by user policies (thus avoiding the need for recomputation).

The way this is written assumes that we can write tags with a particular
user's principal refinement, or implicitly say it applies to all
principals by omitting the principal refinement as proposed by @aferr in
PR #187. This feature does not exist yet, but something like it is
probably necessary to allow the smartmic to express that it downgrades
all userSpeech to speechToText, regardless of who owns it, while also
allowing individual user policies to attach to variants of the tag.