Make multimic example policy check.
This commit runs the policy_check rule upon the multimic example and
alters the multimic example so that it passes and fails the policy check
as expected. The multimic example now has a single graph and multiple
policy files, showing that the example fails and passes as expected in
a number of situations. Also, this alters the policy_check macro to add
an expect_failure parameter to accomodate this use case.

The current state of the multimic example is not perfect. While it would
be nice to show tags delegated to principals through `canSay hasTag`,
that is not working at the moment (see issue #195). As a workaround, we
make every principal that we wish to claim a tag an owner of that tag.

In addition, we are currently modeling the multimic situation as the mic
having the tags userSpeechCanBeDeclassifiedAsTextTranscript and
userSpeechCannotBeDeclassifiedAsTextTranscript. These are intended to be
owned by the microphone, which then allows the users to taint the audio
stream with them to express their preferences. This is kind of abusing
the taint dataflow do some of the authorization logic-y stuff, but
expressing this in a more natural way would probably require
either principal-refined tags or flow decisions based upon tags.