Multimic scenario, strongly tying policies to graph fragments. (#188)
* How we might accomplish multimic example with manifest/policy.

Using the Arcs manifest format and policy format, this attempts to
encode how we might accomplish the example of a user with a microphone
that can

* export its text to speech model to a microphone closer to the user

* detect when a user comes in with an incompatible policy and reason
  about whether to keep recording or stop the text to speech.

* Flesh out multimic a bit more, fix issues in comments.

This commit addresses in a bit more detail the distinct permissions that
might be attached to AudioData and ASRData, details of how UserC ends up
having a stake in the userSpeech tag, and how permissions to access
mic1's storage might be handled.

* Draft of multimic example with policy in authlogic language.

This is a draft of the multimic example that tries to express the policy
enforcement preferences in the authorization logic language. This does
not currently build, as I'm not 100% sure how to express what I want in
the authorization logic language, so putting up a draft for some help.

* Make multimic2 compile

This commit gets multimic2 to compile. It tries to change as little else
about it as possible other than to get it to compile. (It's not tested
in any other way).

Co-authored-by:markww

* Multimic scenario, strongly tying policies to graph fragments.

Following a conversation with @bgogul, this attempt at the multimic
policy attempts to avoid expensive re-analysis of graph properties and
dynamic enforcement of policy by assuming that policies are strongly
associated with a particular graph fragment, within which the
connections cannot change dynamically, and that policies can speak only
of properties within those graph fragments.

This attempt introduces user permission stubs to the multimic graph.
These permission stubs represent the smart mic attempting to accomodate
user policies by describing the speech data as flowing through the
smartphones of guest users. Of course, the actual data does no such
thing, but this allows the smartphones to add tags to the dataflow while
also restricting the particles that they can control to the particles
that they own, allowing the microphone's actual graph to not be altered
by user policies (thus avoiding the need for recomputation).

The way this is written assumes that we can write tags with a particular
user's principal refinement, or implicitly say it applies to all
principals by omitting the principal refinement as proposed by @aferr in
PR #187. This feature does not exist yet, but something like it is
probably necessary to allow the smartmic to express that it downgrades
all userSpeech to speechToText, regardless of who owns it, while also
allowing individual user policies to attach to variants of the tag.

* Fix comment.

* Make multimic example policy check.

This commit runs the policy_check rule upon the multimic example and
alters the multimic example so that it passes and fails the policy check
as expected. The multimic example now has a single graph and multiple
policy files, showing that the example fails and passes as expected in
a number of situations. Also, this alters the policy_check macro to add
an expect_failure parameter to accomodate this use case.

The current state of the multimic example is not perfect. While it would
be nice to show tags delegated to principals through `canSay hasTag`,
that is not working at the moment (see issue #195). As a workaround, we
make every principal that we wish to claim a tag an owner of that tag.

In addition, we are currently modeling the multimic situation as the mic
having the tags userSpeechCanBeDeclassifiedAsTextTranscript and
userSpeechCannotBeDeclassifiedAsTextTranscript. These are intended to be
owned by the microphone, which then allows the users to taint the audio
stream with them to express their preferences. This is kind of abusing
the taint dataflow do some of the authorization logic-y stuff, but
expressing this in a more natural way would probably require
either principal-refined tags or flow decisions based upon tags.

* Improve formatting in multimic BUILD file.

* Remove extraneous pieces of example.

* Remove unnecessary derives from claim and an outdated comment.

Co-authored-by: Andrew Ferraiuolo <aferr@google.com>