Fix delegation with says_canSay_hasTag and says_canSay_downgrades. (#198)
* Fix delegation with says_canSay_hasTag and says_canSay_downgrades.

This PR fixes delegation with says_canSay_hasTag and
says_canSay_downgrades does so by introducing equivalent relations in
authorization_logic.dl (saysCanSayHasTag and saysCanSayDowngrades) and
using them to populate canSayHasTag and canSayDowngrades with the owner
and the owner's delegatees. We add two example test to the examples
directory, simple_canSayDowngrades and simple_canSayHasTag showing this
delegation in action.

In order to get this to work, we also had to work around our current
issues with non-universal perspectives upon universes (issue #184). We
do so by putting a fixed list of known-universe relations within
`lowering_ast_datalog.rs`'s `push_onto_pred` function and essentially
making the function do nothing for those relations, causing them to be
emitted in their universal form rather than be qualified with a `says_`
name prefix and a speaker argument. While this is a bit of a hack, it
does cause the universes to be emitted correctly and only applies to
predicate pieces (the RHS of auth logic datalog statements) so it may be
a good-enough solution.

* Fix comment.

* Add an additional case showing downgrading by non-owner.