Make ir package no longer depend on Arcs proto library. (#203)
* Move proto decoding out of ir::DerivesFromClaim into proto package.

Part of the continuing effort to detangle the arcs protos from the
Raksha ir data structures as described in issue #118.

* Extract Arcs proto decoding from ir::TagClaim.

Part of the ongoing effort to detangle Arcs protos from the Raksha IR
(issue #118).

* Add copyright comment to tag_claim proto files.

* Make ir package no longer depend on Arcs proto library.

This is a small cleanup PR that makes the ir package no longer depend
upon the Arcs protos. It should be merged only after PRs #201 and #202.