Move HandleConnectionSpec into :ir, split proto fns into ir:proto. (#205)
* Move proto decoding out of ir::DerivesFromClaim into proto package.

Part of the continuing effort to detangle the arcs protos from the
Raksha ir data structures as described in issue #118.

* Extract Arcs proto decoding from ir::TagClaim.

Part of the ongoing effort to detangle Arcs protos from the Raksha IR
(issue #118).

* Add copyright comment to tag_claim proto files.

* Make ir package no longer depend on Arcs proto library.

This is a small cleanup PR that makes the ir package no longer depend
upon the Arcs protos. It should be merged only after PRs #201 and #202.

* Remove dependency of ir::types on ir::proto.

This is mostly just deleting some no-longer-used headers, although it
did expose one missing header file in ir/proto/primitive_type.cc . This
is part of the ongoing effort to fix issue #118.

* Move HandleConnectionSpec into :ir, split proto fns into ir:proto.

Prior to this PR, HandleConnectionSpec and ParticleSpec were placed in
//src/xform_to_datalog/arcs_manifest_tree. It has become more clear as
we have worked on the IR that these classes are part of the general
Raksha IR rather than being special to the xform_to_datalog tool, so
I have begun moving them to the IR package.

This PR moves the HandleConnectionSpec class to ir. As part of this
move, I have also split out the proto functions into a separate proto
library. I will handle ParticleSpec in a follow-on PR.

* Fix include guards, namespace comments to match new locatation.

* Add missing copyright notifications.

* Removed extra whitespace around getters.