Untangle Predicate arena from Predicate decoder.
This PR fixes a nasty little tangle between the IR and the protos. Due to
our desire to copy Predicates between spec and impl version of TagChecks,
we place them in an arena to prevent from having to use std::uniqute_ptr
(which restricts copying) while still ensuring that the predicates lived
long enough to be printed out as datalog. The initial way I did this was
by having the PredicateDecoder act as the arena. This is fine if the
only way you are constructing predicates is by parsing protos, but bad
otherwise.

This PR separates the concepts of the PredicateDecoder from the
PredicateArena. It also simplifies the ownership model for Predicates:
while the PredicateArena initially held only the root Predicate pointed
to by a TagCheck, with Predicates pointing to their sub-Predicates with
std::unique_ptrs, now all Predicates are owned by the PredicateArena and
pointers between Predicates are just regular pointers.

To ensure that Predicates are constructed only on arenas, I have made
their constructors private and given them a public static Create factory
method which takes a PredicateArena as an argument. To allow const
global Predicates to be constructed in tests (handy for test
parameterization) I have created a FakePredicateArena which just leaks
the std::unique_ptr it is handed rather than storing it. This allows
creating a global arena for tests that is effectively const due to its
lack of internal state.