Replace FakePredicateArena with SingleUseArenaAndPredicate.
This responds to @bgogul's objections about leaking in the
FakePredicateArena by replacing it with a different test solution.
Instead, we create a SingleUseArenaAndPredicate class, in which is
embeddded an arena and a pointer to a single predicate allocated with
that arena. This allows the arena to be modified only during the
construction of the predicate, preventing having to take in an arena
from the outside. This allows constructing SingleUseArenaAndPredicates
at static scope without requiring a mutable arena to be constructed
prior to it at static scope.

Like physical single-use items, this convenience comes at the cost of
wastefulness. As such, this class should only be used in tests and is
thus placed in its own testonly library.