Untangle Predicate arena from Predicate decoder.  (#206)
* Move proto decoding out of ir::DerivesFromClaim into proto package.

Part of the continuing effort to detangle the arcs protos from the
Raksha ir data structures as described in issue #118.

* Extract Arcs proto decoding from ir::TagClaim.

Part of the ongoing effort to detangle Arcs protos from the Raksha IR
(issue #118).

* Add copyright comment to tag_claim proto files.

* Make ir package no longer depend on Arcs proto library.

This is a small cleanup PR that makes the ir package no longer depend
upon the Arcs protos. It should be merged only after PRs #201 and #202.

* Remove dependency of ir::types on ir::proto.

This is mostly just deleting some no-longer-used headers, although it
did expose one missing header file in ir/proto/primitive_type.cc . This
is part of the ongoing effort to fix issue #118.

* Move HandleConnectionSpec into :ir, split proto fns into ir:proto.

Prior to this PR, HandleConnectionSpec and ParticleSpec were placed in
//src/xform_to_datalog/arcs_manifest_tree. It has become more clear as
we have worked on the IR that these classes are part of the general
Raksha IR rather than being special to the xform_to_datalog tool, so
I have begun moving them to the IR package.

This PR moves the HandleConnectionSpec class to ir. As part of this
move, I have also split out the proto functions into a separate proto
library. I will handle ParticleSpec in a follow-on PR.

* Fix include guards, namespace comments to match new locatation.

* Add missing copyright notifications.

* Untangle Predicate arena from Predicate decoder.

This PR fixes a nasty little tangle between the IR and the protos. Due to
our desire to copy Predicates between spec and impl version of TagChecks,
we place them in an arena to prevent from having to use std::uniqute_ptr
(which restricts copying) while still ensuring that the predicates lived
long enough to be printed out as datalog. The initial way I did this was
by having the PredicateDecoder act as the arena. This is fine if the
only way you are constructing predicates is by parsing protos, but bad
otherwise.

This PR separates the concepts of the PredicateDecoder from the
PredicateArena. It also simplifies the ownership model for Predicates:
while the PredicateArena initially held only the root Predicate pointed
to by a TagCheck, with Predicates pointing to their sub-Predicates with
std::unique_ptrs, now all Predicates are owned by the PredicateArena and
pointers between Predicates are just regular pointers.

To ensure that Predicates are constructed only on arenas, I have made
their constructors private and given them a public static Create factory
method which takes a PredicateArena as an argument. To allow const
global Predicates to be constructed in tests (handy for test
parameterization) I have created a FakePredicateArena which just leaks
the std::unique_ptr it is handed rather than storing it. This allows
creating a global arena for tests that is effectively const due to its
lack of internal state.

* Fix some comments in PredicateArena.

* Replace FakePredicateArena with SingleUseArenaAndPredicate.

This responds to @bgogul's objections about leaking in the
FakePredicateArena by replacing it with a different test solution.
Instead, we create a SingleUseArenaAndPredicate class, in which is
embeddded an arena and a pointer to a single predicate allocated with
that arena. This allows the arena to be modified only during the
construction of the predicate, preventing having to take in an arena
from the outside. This allows constructing SingleUseArenaAndPredicates
at static scope without requiring a mutable arena to be constructed
prior to it at static scope.

Like physical single-use items, this convenience comes at the cost of
wastefulness. As such, this class should only be used in tests and is
thus placed in its own testonly library.