Implement may/will check (#209)
This is meant to implement the may/will check that was previously
described in comments.

Another potentially useful part of this that is not implemented yet is
to throw an error when a particle reads from a handle that has at least
one tag but does not describe how this handle is used with "will". To
implement this, we would need a notion of a particle in the IR and in
the datalog, so this has been put off.