Early start at refactoring AccessPath and AccessPathSelectorSet.
Rather than having an AccessPath be a root combined with a list of
selectors and an AccessPathSelectorSet be a straightforward set of
those AccessPaths, this PR reimagines an AccessPathSelectorSet to be
a tree of selectors with both parent->child pointers and child->parent
pointers. Now, an AccessPath can be viewed as a pointer to a particular
leaf within a tree, and can be printed out by recursing from the leaf to
the root and printing on return. This means that AccessPaths are either
a single pointer or a pair of an AccessPathRoot and a single pointer.

I didn't figure out exactly where these AccessPathSelectorSets should
live (probably somewhere long lived so that we don't have use after
frees when following the pointers in AccessPaths) nor many other
details. But I think this will make copies of AccessPaths less painful
and will make AccessPaths take up a lot less space.