Typing interface
This fixes the typing interface between authorization logic and the
raksha taint analysis and IR. As a result, delegation handled by auth
logic no longer needs to be duplicated in raksha's taint analysis. The
tests for delegation in the combination of raksha and auth logic still
pass as of this PR.

This will conflict with
[#215](https://github.com/google-research/raksha/pull/215), so I will
likely need to redo this later. Even after 215, I think this fix is
still needed in order to handle cases like the following, where in the
authorization logic there is a user-defined relation on the RHS that is used to
prove one of the facts beginning with "says" in Raksha:

```
"P1" saysHasTag("some.path", "someTag") :-
    someUserDefinedRelation("some.path", "someTag").
```

which will cause the declaration:

```
.decl speaker_someUserDefinedRelation(x1: symbol, x2: symbol,
    x3: symbol)
```