Add ownership directly to tags and access paths. (#215)
Various discussions among @bgogul, @aferr and @markww led to the realization that `ownsTag` does not capture enough contextual information for making authoritative decisions at sensitive operations (like `removeTag` and `claimNotEdge`).  This PR distills the various ideas that were discussed into a working prototype. Specifically, this introduces three key concepts:
  - Attach principal to tags instead of claiming it thorough `ownsTag`. 
  - Resolved edge has a principal. 
  - Add the concept of ownership for access paths.

Note: Introducing these concepts has eliminated the need for `edgeToMidpointAccessPath` and explicit rules for `saysCanSayHasTag` in the authorization_logic.dl. Here are the key files of this PR to focus on:

- [taint.dl](https://github.com/google-research/raksha/pull/215/files#diff-c9867611ec9983acdad10ad10d9544af6c55b10b57bdaecdca62ace1e24db56b)
- [authorization_logic.dl](https://github.com/google-research/raksha/pull/215/files#diff-c9668b16b74dbe509b66025a403836b9bce0c6677e2195093b69bedfb00bd0b6)
- [ok_claim_propagates_downgrades_granted.authlogic](https://github.com/google-research/raksha/pull/215/files#diff-a0d4a360467ff8afcc8855d32a309a47a68abb0a326ab2149a6a99ca565d3ef5) for an example use case.

Note that all tests don't pass yet.

 ### Tags have principals.

Instead of having tag to be just a string like `userSelection`, we also associate a principal with the tag and propagate it through the analysis. That is, instead of  `hasTag("a.b.c", "tag")` we will have `hasTag("a.b.c",  "principal", "tag")`. The `"principal"` is propagated throughout the analysis.  (The tag could be implemented as an ADT also if needed.)

This allows us to make authoritative decisions based on the principal for  sensitive operations. e.g.,
```
mayHaveTag(tgt, owner, tag) :-
    resolvedEdge(src, tgt), mayHaveTag(src, owner, tag), !removeTag(tgt, owner, tag).
```

This also allows us to delegate at a much more granular level while not going all the way to something like `speaksFor`. e.g., 

```
// user trusting EFF to handle the tags of "EndUser".
"EndUser" says "EFF" canSay removeTag(path, "EndUser", tag) :- isAccessPath(path), isTag(tag).
// EFF auditing "P1" and making a statement that "P1" removing the "userSelection" tag is valid. 
"EFF" says "P1" canSay removeTag(path, user, "userSelection") :- isAccessPath(path), isPrincipal(user).
```

### Resolved edge has a principal.
Instead of `resolvedEdge(src, tgt)`, we have `resolvedEdge(principal, src, tgt)`, which is read as `src -> tgt` is resolved for principal: 
```
resolvedEdge(principal, src, tgt) :-
   isPrincipal(principal),
   edge(src, tgt),
   !claimNotEdge(principal, src, tgt).
```

### Attaching ownership to access paths. 
This PR also introduces an `ownsAccessPath(principal: Principal, path: AccessPath)` to capture who owns a particular piece of data and is allowed to make decisions operations allowed on the access path. Ideally, we will not be making these claims at the level of access paths within particles. Rather, we will make claims at the point of ingestion of data into the system, which usually corresponds to the handles/stores of a recipe. This ownership of access path is also propagated along the edges unless the edge was removed authoritatively for a principal.
```
ownsAccessPath(principal, tgt) :-
  resolvedEdge(principal, src, tgt), ownsAccessPath(principal, src).
```


* Fix ok_claim_propagages

* Eliminate need for edgeToMidpointAccessPath.

* Fixed all the tests.

* Add grounding rules for handling negation in predicates.

* Handling check with correct ownership of tag. (incomplete).

* Do not use implication for check.

* multi-mic fixes (incomplete).

* Cleaned up further and fixed all tests.

* Remove ToGroundingRules

* Further cleanup.

* Even further cleanup.

* Incorporate changes for may will.

* Added reference to bug.