Add datalog rules for integrity tags.
This PR adds rules for adding, removing, and propagating integrity tags.
This is a little more complex than regular tags, as integrity tags are
only propagated to an AccessPath if all edges pointing to that
AccessPath come from AccessPaths having that integrity tag. This
intersection property cannot be directly expressed in Datalog. Instead,
we propagate the lack of an integrity tag, which we call an inverse
integrity tag, or an InvIntegTag. Propagating the InvIntegTag allows us
to use union operations (ie, a destination access path has an
InvIntegTag if any of its incoming edges have an InvIntegTag) which is
easy to express in Datalog.