Fixes to get C++ Raksha working locally after a toolchain change. (#250)
* Fixes to get C++ Raksha working locally after a toolchain change.

Some piece(s) of my toolchain got updated in the night and broke my
build. These changes seem get the build fixed, at least for the C++ side
which is in my inner loop. The Rust side is still broken for me, but
I don't usually touch that code, so that's less critical for me to fix.

* Add comment about removal of -Werror from host_cxxopts.