Create CheckPredicate for performing checks via facts in DL.
Until now, we've been translating Arcs checks into Datalog rules. While
this gives us a great deal of expressive power, it is tricky to reason
about and requires recompiling before analyzing any Arcs system.

This PR takes the first step in a different direction: instead,
represent checks as regular old facts, with the predicates being
expressed as an ADT. These predicates can then be evaluated inside
Datalog through rules that interpret the predicates encoded in the ADT
rather than encoding the predicates directly as rule bodies.

This PR leaves the infrastructure for the previous way of handling
checks and predicates intact. However, this solution could (and quite
probably should) be used to switch off of that method entirely, allowing
us to precompile the Datalog analysis.