Separate production and test-supporting check fact relations.
In earlier commits, `checkP` fed into `checkAndExpectation`, with all
entries in `checkP` having an expectation of `PASS`. This, however, made
the production `checkP` feed into the test-only `checkAndExpectation`,
which is bad form. This commit refactors that relationship, making
test-supporting relations such as `checkAndExpectation` populate and
read from the production relations such as `checkP`. The test relations
are conditionally compiled, and thus not even present in production
builds.

This change also refactors some naming of intermeidate targets that
participate in a workflow that generates C++ code.