Create CheckPredicate for performing checks via facts in DL. (#263)
* Create CheckPredicate for performing checks via facts in DL.

Until now, we've been translating Arcs checks into Datalog rules. While
this gives us a great deal of expressive power, it is tricky to reason
about and requires recompiling before analyzing any Arcs system.

This PR takes the first step in a different direction: instead,
represent checks as regular old facts, with the predicates being
expressed as an ADT. These predicates can then be evaluated inside
Datalog through rules that interpret the predicates encoded in the ADT
rather than encoding the predicates directly as rule bodies.

This PR leaves the infrastructure for the previous way of handling
checks and predicates intact. However, this solution could (and quite
probably should) be used to switch off of that method entirely, allowing
us to precompile the Datalog analysis.

* Get rid of an unnecessary relation in evaluating predicates.

* Separate production and test-supporting check fact relations.

In earlier commits, `checkP` fed into `checkAndExpectation`, with all
entries in `checkP` having an expectation of `PASS`. This, however, made
the production `checkP` feed into the test-only `checkAndExpectation`,
which is bad form. This commit refactors that relationship, making
test-supporting relations such as `checkAndExpectation` populate and
read from the production relations such as `checkP`. The test relations
are conditionally compiled, and thus not even present in production
builds.

This change also refactors some naming of intermeidate targets that
participate in a workflow that generates C++ code.

* Document behavior of CP_Not more clearly.