Made operator.h use `Block`'s full defn rather than a forward decl. (#295)
* Made operator.h use `Block`'s full defn rather than a forward decl.

The `Operator` class contains within it a member, `implementation_`,
that uses a `std::unique_ptr` pointing to `Block`. Previously, this
header had only a forward declaration for `Block`. This is bad, because
`unique_ptr` needs `Block`'s full definition to get its dtor and to
perform `sizeof` upon that class. This change makes `operator.h` include
the full `block.h` header, giving it access to the full definition of
`Block`.

* Get rid of unnecessary headers in data_decl.h