Ue raw string to enable substitutions in string extractor. (#343)
Make the string that prints the include in dl_string_test_file_generator
be a raw string instead of an escaped string. Previously, we kept the
string in dl_string_test_file_generator as an escaped string rather than
a raw string because it was very concise. The escaping, however,
prevents us from automatically re-rooting this include when copying into
Google's internal repo.