Make IRContext manage Storages in addition to Operators.
Previously, IRContext managed the mapping of names to Operators. This
change makes it also manage the mapping of names to Storages and keep
track of canonical copies of Operations. Because the logic for Operators
and Storages is so similar, an inner templatized class is used to
serve both kinds of IrNodes.