Make IRContext manage Storages in addition to Operators. (#385)
* Make attributes and values of operations be printed out in name order.

Previously, the attributes and names of operations would be printed out
in arbitrary order. As of this commit, they are always printed out in
the alphabetical order of their associated name.

* Make IRContext manage Storages in addition to Operators.

Previously, IRContext managed the mapping of names to Operators. This
change makes it also manage the mapping of names to Storages and keep
track of canonical copies of Operations. Because the logic for Operators
and Storages is so similar, an inner templatized class is used to
serve both kinds of IrNodes.

* Rename Ir -> IR in all names.

* Shorten type parameter from NamedIRNode to Node.

* Back out some changes enabling move on `Operation` that are not needed.

* Simplify interface by moving node kind name into template argument.