Add protos used for communicating with SQL verifier for MVP impl.
Add the protos that we will use to communicate with the SQL verifier
for the MVP. These protos are built to make it very easy to export the
minimum amount of state possible from the SQL verifier to provide some
dataflow analysis value. A mature implementation should probably throw
these protos out entirely and rewrite them to capture more information
and be better factored.