Decode `Literal` SQL IR protos.
This PR decodes `Literal` SQL IR protos. `Literal` protos are decoded
almost identically to `SourceTableColumn`s, even down to representing
them as `Storage`s. This is done because the IR does not currently have
a representation for `Storage`s and because in a verifier that does not
allow for modifications (as with the SQL verifier we are targeting),
there actually isn't that much difference between a named `Storage` and
a `Literal` associated with some canonical name.