Change `TagTransform`'s map to allow multiple tags on a single expr.
Previously, `TagTransform` had a first-class proto `map` indicating its
tag preconditions. This precluded a single expression being required to
have multiple `IntegrityTag`s, which seemed perfectly possible. So we've
changed the "map" to instead be a list of pairs of expression ID and
required integrity tag to handle this case.