Add protos used for communicating with SQL verifier for MVP impl. (#389)
* Add protos used for communicating with SQL verifier for MVP impl.

Add the protos that we will use to communicate with the SQL verifier
for the MVP. These protos are built to make it very easy to export the
minimum amount of state possible from the SQL verifier to provide some
dataflow analysis value. A mature implementation should probably throw
these protos out entirely and rewrite them to capture more information
and be better factored.

* Buildifier fixup.

* Refactor SQL protos to emphasize `Expression` superclass.

This change refactors the SQL protos to put more logic into and to
emphasize the `Expression` (previously `Value`) message as a single
superclass for all other SQL messages. Each `Expression` has a
required `id` and optional `name`. `PolicyTransform`s act as casts upon
some `Expression`, and may add or remove either an integrity or
confidentiality tag. The list of tags on `SourceTableColumn`s has been
removed; they will now be realized with applications of
`PolicyTransform`s.

* Rename PolicyTransfrom -> TagTransform.

* Change `TagTransform`'s map to allow multiple tags on a single expr.

Previously, `TagTransform` had a first-class proto `map` indicating its
tag preconditions. This precluded a single expression being required to
have multiple `IntegrityTag`s, which seemed perfectly possible. So we've
changed the "map" to instead be a list of pairs of expression ID and
required integrity tag to handle this case.

* Clarify behavior of `MergeOperation` with a comment.