Move `GetOrCreateStorage` to new `DecoderContext` class.
@bgogul pointed out that we don't want to lazily create `Operator`s and
`Storage`s in most cases. For the SQL MVP, however, it is very useful to
lazily create `Storage`s because of how SQL buries its context at the
bottom of the queries. This PR creates a new `DecoderContext` to keep
state and methods that are SQL specific, and adds a `GetOrCreateStorage`
there.