Decode `Literal` SQL IR protos. (#392)
* Add protos used for communicating with SQL verifier for MVP impl.

Add the protos that we will use to communicate with the SQL verifier
for the MVP. These protos are built to make it very easy to export the
minimum amount of state possible from the SQL verifier to provide some
dataflow analysis value. A mature implementation should probably throw
these protos out entirely and rewrite them to capture more information
and be better factored.

* Buildifier fixup.

* Refactor SQL protos to emphasize `Expression` superclass.

This change refactors the SQL protos to put more logic into and to
emphasize the `Expression` (previously `Value`) message as a single
superclass for all other SQL messages. Each `Expression` has a
required `id` and optional `name`. `PolicyTransform`s act as casts upon
some `Expression`, and may add or remove either an integrity or
confidentiality tag. The list of tags on `SourceTableColumn`s has been
removed; they will now be realized with applications of
`PolicyTransform`s.

* Add `GetOrCreateNode` to `IRContext`.

This change adds `GetOrCreateNode` to `IRContext`. This allows creating
an entry if it does not already exist, returning the one that does exist
otherwise. This is just implemented as an `insert` in the underlying
map.

* Decode SourceTableColumn SQL proto.

This PR decodes the SourceTableColumn SQL proto. The decoding of this
proto is fairly straightforward, as it just involves getting the storage
indicated by the proto (creating it if it does not yet exist) and
wrapping the whole thing in a `StorageValue` and a `Value`.

* Decode `Literal` SQL IR protos.

This PR decodes `Literal` SQL IR protos. `Literal` protos are decoded
almost identically to `SourceTableColumn`s, even down to representing
them as `Storage`s. This is done because the IR does not currently have
a representation for `Storage`s and because in a verifier that does not
allow for modifications (as with the SQL verifier we are targeting),
there actually isn't that much difference between a named `Storage` and
a `Literal` associated with some canonical name.

* Move `GetOrCreateStorage` to new `DecoderContext` class.

@bgogul pointed out that we don't want to lazily create `Operator`s and
`Storage`s in most cases. For the SQL MVP, however, it is very useful to
lazily create `Storage`s because of how SQL buries its context at the
bottom of the queries. This PR creates a new `DecoderContext` to keep
state and methods that are SQL specific, and adds a `GetOrCreateStorage`
there.

* Fix up uses of GetOrCreateStorage now that it is on DecoderContext.

* Add missing dependency from `decode` to `decoder_context`.

* Fix now that `GetOrCreateStorage` is on `DecoderContext`.

* Represent `Literal`s with an `Operation` rather than a `Storage`.

* Create `constexpr` globals for constant names used in SQL decoding.

* Refactor tests to improve readability.

* Fix `ResultOf` matcher by using lambdas rather than pointer-to-members.